Refactor HomePage by splitting into a separate component
